third_party/blink/web_tests/fast/dom/Window/window-property-invalid-characters-ignored.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script>
jsTestIsAsync = true;

if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.setPopupBlockingEnabled(false);
}

description("Tests that Unicode characters that changes the string length when lower cased still works.");

// U+0130 = LATIN CAPITAL LETTER I WITH DOT ABOVE
// U+0069 LATIN SMALL LETTER I
// U+0307 COMBINING DOT ABOVE
// \u0130 gets lowered to \u0069\u0307.

var w = window.open("resources/window-property-invalid-characters-ignored.html", "blank", "\u0130\u0130=\u0130\u0130,width=123,height=123");

function finishTest()
{
    shouldBe("w.innerWidth", "123");
    shouldBe("w.innerHeight", "123");
    shouldBeTrue("w !== window");
    w.close();

    finishJSTest();
}

</script>
</head>
<body>
<pre id="console"></pre>
</body>
</html>
