third_party/blink/web_tests/platform/generic/fast/dom/Window/window-property-invalid-characters-ignored-expected.txt

Tests that Unicode characters that changes the string length when lower cased still works.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS w.innerWidth is 123
PASS w.innerHeight is 123
PASS w !== window is true
PASS successfullyParsed is true

TEST COMPLETE
